+++ packages/open-agents/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @davstack/open-agents
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @davstack/cli-utils@1.3.0
+
 ## 1.2.2
 
 ### Patch Changes